Refactor: Enhance DB Update Logic and Timer Service
This commit introduces a more robust database update mechanism that checks remote JSON versions against local cached versions and improves the TimerService behavior.

**Database Update Enhancements:**
- **`PopulateDatabaseCallback.kt`:**
    - `checkAndPopulateDatabase`:
        - Now fetches remote JSON version first.
        - Compares remote version with the local version stored in SharedPreferences (per language).
        - If remote is newer, downloads and saves the new JSON to local cache, then updates the database.
        - If remote is not newer or inaccessible, uses local cached JSON if available.
        - `ExercisesVersion.version` changed from `Int` to `Double` to support more granular versioning.
    - `populateFromJson`: New function to handle DB population from a given JSON string, allowing for forced updates.
    - `downloadJsonFromGitHub`: Made public for use by `ExerRepositoryImpl`.
    - Removed `onCreate` and `onOpen` overrides as DB population is now handled more explicitly.
    - Removed `prefetchAllLanguages` as updates are checked on demand.
- **`ExerRepository.kt` & `ExerRepositoryImpl.kt`:**
    - Added `checkForUpdates(lang: String): Boolean` to check for and apply updates, returning `true` if an update occurred.
- **`HomeViewModel.kt`:**
    - Added `checkForUpdatesSuspend(lang: String): Boolean` to call the repository's update check.
    - Handles `HomeEvent.CheckUpdates` to trigger the update check and show a `Snackbar` with the result via `UiEvent.ShowSnackbar`.
- **`Home.kt`:**
    - Added `SnackbarHost` to display update messages.
    - The "Refresh" button in `HomeBottomBar` now triggers `HomeEvent.CheckUpdates`.
    - `coroutineScope` used to launch the update check on `Dispatchers.IO`.
- **`HomeEvent.kt`:**
    - Added `CheckUpdates` event.
- **`UiEvent.kt`:**
    - Added `ShowSnackbar(val message: String)` event.
- **`ExerDAO.kt`:**
    - (Indirectly affected by `PopulateDatabaseCallback` changes, ensuring `clearExercises` and `insertAll` are used during updates).
- **`ExercisesVersion.kt`:**
    - Changed `version` type from `Int` to `Double`.

**Timer Service & UI Improvements:**
- **`TimerService.kt`:**
    - `increment()`: Now correctly stops and resumes the timer if running, or just updates times if paused.
    - `decrement()`: Now correctly stops and resumes the timer if running, or just updates times if paused; ensures time doesn't go below zero.
    - `start()`/`resume()` now call `toForeground()` to ensure service stays active.
    - `stop()`/`reset()` now call `toBackground()`.
    - `WorkoutTimer.onTick()`: Removed redundant check for `time <= 0L` as `onFinish()` is called by the system.
    - `Long.toTimerString()`: Formats seconds with leading zero (`%02d`).
- **`TimerBar.kt`:**
    - `Long.toTimerString()` (local version) updated for correct formatting (`%d:%02d`).
- **`SessionBottomBar.kt`:**
    - Simplified FAB `onClick` for timer to directly call `onTimerPress`.
    - `TimerAction`'s `onClick` now directly uses `onTimerPress`.
    - Corrected FAB size modifiers using `Modifier.size()` and `Modifier.width().height()`.
- **`TimerAction.kt`:**
    - Simplified `IconButton` `onClick`.
- **`AndroidManifest.xml`:**
    - Added `FOREGROUND_SERVICE_DATA_SYNC` and `FOREGROUND_SERVICE` permissions.

**Minor Changes:**
- **`Settings.kt`:**
    - Added "TODO:" prefix to "Export Database" and "Import Database" button texts.
- **`Session.kt`:**
    - Added `UiEvent.ShowSnackbar` to the `when` block in `LaunchedEffect` (though not used yet).

**Impact:**
- Database updates are now more reliable, checking versions before applying and using cached data when offline.
- Users receive feedback on database update status via Snackbars.
- TimerService behavior is more consistent, especially when incrementing/decrementing while running or paused.
- Timer display format is corrected.
- Ensures TimerService uses foreground capabilities correctly, improving reliability.

+++ app/src/main/java/com/example/exerlog/db/entities/ExercisesVersion.kt
@@ -1,6 +1,6 @@
 package com.example.exerlog.db.entities
 
 data class ExercisesVersion(
-    val version: Int,
+    val version: Double,
     val exercises: List<Exercise>
 )
